extracted ui logic from LabelPickerDialog into LabelPickerUILogic
