+++ src/main/java/ui/components/pickers/LabelPickerDialog.java
@@ -11,8 +11,8 @@ import javafx.scene.layout.VBox;
 import javafx.stage.Modality;
 import javafx.stage.Stage;
 
-import java.util.*;
-import java.util.regex.Pattern;
+import java.util.List;
+import java.util.Map;
 import java.util.stream.Collectors;
 
 public class LabelPickerDialog extends Dialog<List<String>> {
@@ -20,39 +20,24 @@ public class LabelPickerDialog extends Dialog<List<String>> {
     private static final int VBOX_SPACING = 105; // seems like some magic number
     private static final int ELEMENT_MAX_WIDTH = 400;
 
-    private TurboIssue issue;
+    private final LabelPickerUILogic uiLogic;
     private TextField textField;
-    private List<TurboLabel> allLabels;
-    private List<PickerLabel> topLabels = new ArrayList<>();
-    private List<PickerLabel> bottomLabels;
-    private Set<String> groups = new HashSet<>();
-    private Map<String, Boolean> resultList = new HashMap<>();
     private FlowPane topPane;
     private FlowPane bottomPane;
-    private Optional<String> targetLabel = Optional.empty();
 
-    // Used for multiple spaces
-    private String lastAction = "";
-    private int previousNumberOfActions = 0;
-
-    LabelPickerDialog(TurboIssue issue, List<TurboLabel> allLabels, Stage stage) {
-        this.issue = issue;
-        initialiseDialog(stage);
-        populateAllLabels(allLabels);
-
+    LabelPickerDialog(TurboIssue issue, List<TurboLabel> repoLabels, Stage stage) {
         // UI creation
+        initialiseDialog(stage, issue);
         createButtons();
         VBox vBox = createVBox();
-        Label titleLabel = createTitleLabel();
-        titleLabel.setTooltip(createTitleTooltip());
+        Label titleLabel = createTitleLabel(issue);
+        titleLabel.setTooltip(createTitleTooltip(issue));
         topPane = createTopPane();
         textField = createTextField();
         bottomPane = createBottomPane();
 
         setupKeyEvents();
-        addExistingLabels();
-        updateBottomLabels("");
-        populatePanes();
+        uiLogic = new LabelPickerUILogic(issue, repoLabels, this);
 
         vBox.getChildren().addAll(titleLabel, topPane, textField, bottomPane);
         getDialogPane().setContent(vBox);
@@ -62,319 +47,32 @@ public class LabelPickerDialog extends Dialog<List<String>> {
 
     private void setupKeyEvents() {
         textField.textProperty().addListener((observable, oldValue, newValue) -> {
-            processTextFieldChange(newValue.toLowerCase());
+            uiLogic.processTextFieldChange(newValue.toLowerCase());
         });
         textField.setOnKeyPressed(e -> {
             if (!e.isAltDown() && !e.isMetaDown() && !e.isControlDown()) {
                 if (e.getCode() == KeyCode.DOWN) {
                     e.consume();
-                    moveHighlightOnLabel(true);
+                    uiLogic.moveHighlightOnLabel(true);
                 } else if (e.getCode() == KeyCode.UP) {
                     e.consume();
-                    moveHighlightOnLabel(false);
+                    uiLogic.moveHighlightOnLabel(false);
                 } else if (e.getCode() == KeyCode.SPACE) {
                     e.consume();
-                    toggleSelectedLabel();
+                    uiLogic.toggleSelectedLabel(textField.getText());
                 }
             }
         });
     }
 
-    public void toggleLabel(String name) {
-        preProcessAndUpdateTopLabels(name);
-        updateBottomLabels(""); // clears search query, removes faded-out overlay on bottom labels
-        populatePanes();
-    }
-
-    private void toggleSelectedLabel() {
-        if (!bottomLabels.isEmpty() && !textField.getText().isEmpty() && hasHighlightedLabel()) {
-            toggleLabel(
-                    bottomLabels.stream().filter(PickerLabel::isHighlighted).findFirst().get().getActualName());
-        }
-    }
-
-    private boolean containsIgnoreCase(String source, String query) {
-        return source.toLowerCase().contains(query.toLowerCase());
-    }
-
-    private void processTextFieldChange(String text) {
-        String[] textArray = text.split(" ");
-        if (textArray.length > 0) {
-            String query = textArray[textArray.length - 1];
-            if (previousNumberOfActions != textArray.length || !query.equals(lastAction)) {
-                previousNumberOfActions = textArray.length;
-                lastAction = query;
-                boolean isBottomLabelsUpdated = false;
-
-                // group check
-                if (TurboLabel.getDelimiter(query).isPresent()) {
-                    String delimiter = TurboLabel.getDelimiter(query).get();
-                    String[] queryArray = query.split(Pattern.quote(delimiter));
-                    if (queryArray.length == 1) {
-                        isBottomLabelsUpdated = true;
-                        updateBottomLabels(queryArray[0], "");
-                    } else if (queryArray.length == 2) {
-                        isBottomLabelsUpdated = true;
-                        updateBottomLabels(queryArray[0], queryArray[1]);
-                    }
-                }
-
-                if (!isBottomLabelsUpdated) {
-                    updateBottomLabels(query);
-                }
-
-                if (hasHighlightedLabel()) {
-                    addRemovePossibleLabel(getHighlightedLabelName().get().getActualName());
-                } else {
-                    addRemovePossibleLabel("");
-                }
-                populatePanes();
-            }
-        }
-    }
-
-    private void ______TOP_PANE______() {}
-
-    private void preProcessAndUpdateTopLabels(String name) {
-        Optional<TurboLabel> turboLabel =
-                allLabels.stream().filter(label -> label.getActualName().equals(name)).findFirst();
-        if (turboLabel.isPresent()) {
-            if (turboLabel.get().isExclusive() && !resultList.get(name)) {
-                // exclusive label check
-                String group = turboLabel.get().getGroup().get();
-                allLabels
-                        .stream()
-                        .filter(TurboLabel::isExclusive)
-                        .filter(label -> label.getGroup().get().equals(group))
-                        .forEach(label -> updateTopLabels(label.getActualName(), false));
-                updateTopLabels(name, true);
-            } else {
-                updateTopLabels(name, !resultList.get(name));
-            }
-        }
-    }
-
-    private void updateTopLabels(String name, boolean isAdd) {
-        // adds new labels to the end of the list
-        resultList.put(name, isAdd); // update resultList first
-        if (isAdd) {
-            if (issue.getLabels().contains(name)) {
-                topLabels.stream()
-                        .filter(label -> label.getActualName().equals(name))
-                        .forEach(label -> {
-                            label.setIsRemoved(false);
-                            label.setIsFaded(false);
-                        });
-            } else {
-                allLabels.stream()
-                        .filter(label -> label.getActualName().equals(name))
-                        .filter(label -> resultList.get(label.getActualName()))
-                        .filter(label -> !isInTopLabels(label.getActualName()))
-                        .findFirst()
-                        .ifPresent(label -> topLabels.add(new PickerLabel(label, this)));
-            }
-        } else {
-            topLabels.stream()
-                    .filter(label -> label.getActualName().equals(name))
-                    .findFirst()
-                    .ifPresent(label -> {
-                        if (issue.getLabels().contains(name)) {
-                            label.setIsRemoved(true);
-                        } else {
-                            topLabels.remove(label);
-                        }
-                    });
-        }
-    }
-
-    private boolean isInTopLabels(String name) {
-        // used to prevent duplicates in topLabels
-        return topLabels.stream()
-                .filter(label -> label.getActualName().equals(name))
-                .findAny()
-                .isPresent();
-    }
-
-    private void addRemovePossibleLabel(String name) {
-        // Deletes previous selection
-        if (targetLabel.isPresent()) {
-            // if there's a previous possible selection, delete it
-            // targetLabel can be
-            topLabels.stream()
-                    .filter(label -> label.getActualName().equals(targetLabel.get()))
-                    .findFirst()
-                    .ifPresent(label -> {
-                        if (issue.getLabels().contains(targetLabel.get()) ||
-                                resultList.get(targetLabel.get())) {
-                            // if it is an existing label toggle fade and strike through
-                            label.setIsHighlighted(false);
-                            label.setIsFaded(false);
-                            if (resultList.get(label.getActualName())) {
-                                label.setIsRemoved(false);
-                            } else {
-                                label.setIsRemoved(true);
-                            }
-                        } else {
-                            // if not then remove it
-                            topLabels.remove(label);
-                        }
-                    });
-            targetLabel = Optional.empty();
-        }
-
-        if (!name.isEmpty()) {
-            // Try to add current selection
-            if (isInTopLabels(name)) {
-                // if it exists in the top pane
-                topLabels.stream()
-                        .filter(label -> label.getActualName().equals(name))
-                        .findFirst()
-                        .ifPresent(label -> {
-                            label.setIsHighlighted(true);
-                            if (issue.getLabels().contains(name)) {
-                                // if it is an existing label toggle fade and strike through
-                                label.setIsFaded(resultList.get(name));
-                                label.setIsRemoved(resultList.get(name));
-                            } else {
-                                // else set fade and strike through
-                                // if space is pressed afterwards, label is removed from topLabels altogether
-                                label.setIsFaded(true);
-                                label.setIsRemoved(true);
-                            }
-                        });
-            } else {
-                // add it to the top pane
-                allLabels.stream()
-                        .filter(label -> label.getActualName().equals(name))
-                        .findFirst()
-                        .ifPresent(label -> topLabels.add(new PickerLabel(label, this, false, true, false, true)));
-            }
-            targetLabel = Optional.of(name);
-        }
-    }
-
-    private void ______BOTTOM_PANE______() {}
-
-    private void updateBottomLabels(String group, String match) {
-        boolean isValidGroup = groups.stream()
-                .filter(validGroup -> validGroup.startsWith(group))
-                .findAny()
-                .isPresent();
-
-        if (isValidGroup) {
-            List<String> validGroups = groups.stream()
-                    .filter(validGroup -> validGroup.startsWith(group))
-                    .collect(Collectors.toList());
-            // get all labels that contain search query
-            // fade out labels which do not match
-            bottomLabels = allLabels
-                    .stream()
-                    .map(label -> new PickerLabel(label, this))
-                    .map(label -> {
-                        if (resultList.get(label.getActualName())) {
-                            label.setIsSelected(true); // add tick if selected
-                        }
-                        if (!label.getGroup().isPresent() ||
-                                !validGroups.contains(label.getGroup().get()) ||
-                                !containsIgnoreCase(label.getName(), match)) {
-                            label.setIsFaded(true); // fade out if does not match search query
-                        }
-                        return label;
-                    })
-                    .collect(Collectors.toList());
-            if (!bottomLabels.isEmpty()) highlightFirstMatchingItem(match);
-        } else {
-            updateBottomLabels(match);
-        }
-    }
-
-    private void updateBottomLabels(String match) {
-        // get all labels that contain search query
-        // fade out labels which do not match
-        bottomLabels = allLabels
-                .stream()
-                .map(label -> new PickerLabel(label, this))
-                .map(label -> {
-                    if (resultList.get(label.getActualName())) {
-                        label.setIsSelected(true); // add tick if selected
-                    }
-                    if (!match.isEmpty() && !containsIgnoreCase(label.getActualName(), match)) {
-                        label.setIsFaded(true); // fade out if does not match search query
-                    }
-                    return label;
-                })
-                .collect(Collectors.toList());
-
-        if (!match.isEmpty() && !bottomLabels.isEmpty()) highlightFirstMatchingItem(match);
-    }
-
-    private void moveHighlightOnLabel(boolean isDown) {
-        if (hasHighlightedLabel()) {
-            // used to move the highlight on the bottom labels
-            // find all matching labels
-            List<PickerLabel> matchingLabels = bottomLabels.stream()
-                    .filter(label -> !label.isFaded())
-                    .collect(Collectors.toList());
-
-            // move highlight around
-            for (int i = 0; i < matchingLabels.size(); i++) {
-                if (matchingLabels.get(i).isHighlighted()) {
-                    if (isDown && i < matchingLabels.size() - 1) {
-                        matchingLabels.get(i).setIsHighlighted(false);
-                        matchingLabels.get(i + 1).setIsHighlighted(true);
-                        addRemovePossibleLabel(matchingLabels.get(i + 1).getActualName());
-                    } else if (!isDown && i > 0) {
-                        matchingLabels.get(i - 1).setIsHighlighted(true);
-                        matchingLabels.get(i).setIsHighlighted(false);
-                        addRemovePossibleLabel(matchingLabels.get(i - 1).getActualName());
-                    }
-                    populatePanes();
-                    return;
-                }
-            }
-        }
-    }
-
-    private void highlightFirstMatchingItem(String match) {
-        List<PickerLabel> matches = bottomLabels.stream()
-                .filter(label -> !label.isFaded())
-                .collect(Collectors.toList());
-
-        // try to highlight labels that begin with match first
-        matches.stream()
-                .filter(label -> label.getName().startsWith(match))
-                .findFirst()
-                .ifPresent(label -> label.setIsHighlighted(true));
-
-        // if not then highlight first matching label
-        if (!hasHighlightedLabel()) {
-            matches.stream()
-                    .findFirst()
-                    .ifPresent(label -> label.setIsHighlighted(true));
-        }
-    }
-
-    private boolean hasHighlightedLabel() {
-        return bottomLabels.stream()
-                .filter(PickerLabel::isHighlighted)
-                .findAny()
-                .isPresent();
-    }
-
-    private Optional<PickerLabel> getHighlightedLabelName() {
-        return bottomLabels.stream()
-                .filter(PickerLabel::isHighlighted)
-                .findAny();
-    }
-
     private void ______PANE_POPULATION______() {}
 
-    private void populatePanes() {
-        populateTopPane();
-        populateBottomPane();
+    protected void populatePanes(List<PickerLabel> topLabels, List<PickerLabel> bottomLabels) {
+        populateTopPane(topLabels);
+        populateBottomPane(bottomLabels);
     }
 
-    private void populateTopPane() {
+    private void populateTopPane(List<PickerLabel> topLabels) {
         topPane.getChildren().clear();
         topLabels.forEach(label -> topPane.getChildren().add(label.getNode()));
         if (topPane.getChildren().size() == 0) {
@@ -384,7 +82,7 @@ public class LabelPickerDialog extends Dialog<List<String>> {
         }
     }
 
-    private void populateBottomPane() {
+    private void populateBottomPane(List<PickerLabel> bottomLabels) {
         bottomPane.getChildren().clear();
         bottomLabels.forEach(label -> bottomPane.getChildren().add(label.getNode()));
         if (bottomPane.getChildren().size() == 0) {
@@ -394,34 +92,14 @@ public class LabelPickerDialog extends Dialog<List<String>> {
         }
     }
 
-    private void ______INITIALISATION______() {}
+    private void ______UI_CREATION______() {}
 
-    private void initialiseDialog(Stage stage) {
+    private void initialiseDialog(Stage stage, TurboIssue issue) {
         initOwner(stage);
         initModality(Modality.APPLICATION_MODAL); // TODO change to NONE for multiple dialogs
         setTitle("Edit Labels for " + (issue.isPullRequest() ? "PR #" : "Issue #") +
                 issue.getId() + " in " + issue.getRepoId());
     }
-
-    private void populateAllLabels(List<TurboLabel> allLabels) {
-        this.allLabels = new ArrayList<>(allLabels);
-        Collections.sort(this.allLabels);
-        // populate resultList by going through allLabels and seeing which ones currently exist
-        // in issue.getLabels()
-        allLabels.forEach(label -> {
-            resultList.put(label.getActualName(), issue.getLabels().contains(label.getActualName()));
-            if (label.getGroup().isPresent()) groups.add(label.getGroup().get());
-        });
-    }
-
-    private void addExistingLabels() {
-        // used once to populate topLabels at the start
-        allLabels.stream()
-                .filter(label -> issue.getLabels().contains(label.getActualName()))
-                .forEach(label -> topLabels.add(new PickerLabel(label, this)));
-    }
-
-    private void ______UI_CREATION______() {}
 
     private void createButtons() {
         ButtonType confirmButtonType = new ButtonType("Confirm", ButtonBar.ButtonData.OK_DONE);
@@ -431,9 +109,9 @@ public class LabelPickerDialog extends Dialog<List<String>> {
         setResultConverter(dialogButton -> {
             if (dialogButton == confirmButtonType) {
                 // if there is a highlighted label, toggle that label first
-                if (hasHighlightedLabel()) toggleSelectedLabel();
+                if (uiLogic.hasHighlightedLabel()) uiLogic.toggleSelectedLabel(textField.getText());
                 // if user confirms selection, return list of labels
-                return resultList.entrySet().stream()
+                return uiLogic.getResultList().entrySet().stream()
                         .filter(Map.Entry::getValue)
                         .map(Map.Entry::getKey)
                         .collect(Collectors.toList());
@@ -452,7 +130,7 @@ public class LabelPickerDialog extends Dialog<List<String>> {
         return vBox;
     }
 
-    private Label createTitleLabel() {
+    private Label createTitleLabel(TurboIssue issue) {
         Label titleLabel = new Label(
                 (issue.isPullRequest() ? "PR #" : "Issue #") + issue.getId() + ": " + issue.getTitle());
         titleLabel.setMaxWidth(ELEMENT_MAX_WIDTH);
@@ -460,7 +138,7 @@ public class LabelPickerDialog extends Dialog<List<String>> {
         return titleLabel;
     }
 
-    private Tooltip createTitleTooltip() {
+    private Tooltip createTitleTooltip(TurboIssue issue) {
         Tooltip titleTooltip = new Tooltip(
                 (issue.isPullRequest() ? "PR #" : "Issue #") + issue.getId() + ": " + issue.getTitle());
         titleTooltip.setWrapText(true);
