+++ src/main/java/ui/components/pickers/PickerLabel.java
@@ -8,25 +8,25 @@ import javafx.scene.control.Tooltip;
 // for use with LabelPickerDialog
 public class PickerLabel extends TurboLabel {
 
-    private LabelPickerDialog labelPickerDialog;
+    private LabelPickerUILogic labelPickerUILogic;
     private boolean isSelected;
     private boolean isHighlighted;
     private boolean isRemoved;
     private boolean isFaded;
 
-    public PickerLabel(TurboLabel label, LabelPickerDialog labelPickerDialog) {
+    public PickerLabel(TurboLabel label, LabelPickerUILogic labelPickerUILogic) {
         super(label.getRepoId(), label.getColour(), label.getActualName());
-        this.labelPickerDialog = labelPickerDialog;
+        this.labelPickerUILogic = labelPickerUILogic;
         isSelected = false;
         isHighlighted = false;
         isRemoved = false;
         isFaded = false;
     }
 
-    public PickerLabel(TurboLabel label, LabelPickerDialog labelPickerDialog,
+    public PickerLabel(TurboLabel label, LabelPickerUILogic labelPickerUILogic,
                        boolean isSelected, boolean isHighlighted, boolean isRemoved, boolean isFaded) {
         super(label.getRepoId(), label.getColour(), label.getActualName());
-        this.labelPickerDialog = labelPickerDialog;
+        this.labelPickerUILogic = labelPickerUILogic;
         this.isSelected = isSelected;
         this.isHighlighted = isHighlighted;
         this.isRemoved = isRemoved;
@@ -47,7 +47,7 @@ public class PickerLabel extends TurboLabel {
             label.setTooltip(groupTooltip);
         }
 
-        label.setOnMouseClicked(e -> labelPickerDialog.toggleLabel(getActualName()));
+        label.setOnMouseClicked(e -> labelPickerUILogic.toggleLabel(getActualName()));
         return label;
     }
